Garmin: extract the devinfo from the first FIT file we parse
A typical FIT file contains several DEVICE_INFO messages. We need to identify
the one(s) for the creator (i.e. the actual device, not a sub-component).

Note, Garmin identifies the Descent Mk1 as product 2859. I think we should use
this as the model number (instead of currently using 0.

Also, the vendor event is not to send the vendor name of the device, but in
order to send vendor specific events :-)

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>